refactor(model): extract libmlx types to decouple model from rpc
- move MlxDeviceInfo and FirmwareFlashReport from libmlx into new carbide-libmlx-model crate
- relocate protobuf/RPC conversion logic into rpc::libmlx
- update api/model/db/cli consumers to depend on the shared model crate directly
- continue larger refactor to remove model -> rpc coupling so model users do not implicitly depend on RPC
