+++ crates/admin-cli/Cargo.toml
@@ -40,6 +40,7 @@ carbide-secrets = { path = "../secrets" }
 carbide-health-report = { path = "../health-report" }
 carbide-measured-boot = { path = "../measured-boot", features = ["cli"] }
 carbide-libmlx = { path = "../libmlx" }
+carbide-libmlx-model = { path = "../libmlx-model" }
 carbide-rpc = { path = "../rpc", features = ["cli"] }
 carbide-rpc-utils = { path = "../rpc-utils" }
 carbide-utils = { path = "../utils" }
