+++ crates/api-db/src/dpa_interface.rs
@@ -18,11 +18,11 @@
 use std::collections::HashSet;
 use std::net::IpAddr;
 
+use carbide_libmlx_model::device::info::MlxDeviceInfo;
 use carbide_uuid::dpa_interface::{DpaInterfaceId, NULL_DPA_INTERFACE_ID};
 use carbide_uuid::machine::MachineId;
 use config_version::ConfigVersion;
 use eyre::eyre;
-use libmlx::device::info::MlxDeviceInfo;
 use mac_address::MacAddress;
 use model::controller_outcome::PersistentStateHandlerOutcome;
 use model::dpa_interface::{
@@ -561,8 +561,8 @@ pub async fn try_update_network_config(
 mod test {
     use std::str::FromStr;
 
+    use carbide_libmlx_model::device::info::MlxDeviceInfo;
     use carbide_uuid::machine::MachineId;
-    use libmlx::device::info::MlxDeviceInfo;
     use mac_address::MacAddress;
     use model::dpa_interface::NewDpaInterface;
     use model::machine::ManagedHostState;
