+++ crates/libmlx-model/src/firmware/mod.rs
@@ -15,8 +15,4 @@
  * limitations under the License.
  */
 
-pub mod cmd;
-pub mod discovery;
-pub mod filters;
-pub mod proto;
-pub mod report;
+pub mod result;
