+++ crates/libmlx/src/device/cmd/device/cmds.rs
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
+use carbide_libmlx_model::device::info::MlxDeviceInfo;
 use prettytable::{Cell, Row, Table};
 
 use crate::device::cmd::device::args::{DeviceAction, DeviceArgs, OutputFormat};
 use crate::device::filters::{DeviceFilter, DeviceFilterSet};
-use crate::device::info::MlxDeviceInfo;
 use crate::device::report::MlxDeviceReport;
 
 // build_filter_set creates a DeviceFilterSet from a vector of DeviceFilters.
@@ -282,7 +282,7 @@ fn wrap_text(text: &str, width: usize) -> String {
 // resolve_device finds a device by its PCI address or identifier.
 pub fn resolve_device(
     device_spec: &str,
-) -> Result<crate::device::info::MlxDeviceInfo, Box<dyn std::error::Error>> {
+) -> Result<carbide_libmlx_model::device::info::MlxDeviceInfo, Box<dyn std::error::Error>> {
     // Use existing device discovery logic.
     let devices = crate::device::discovery::discover_devices()?;
