+++ crates/libmlx/src/device/discovery.rs
@@ -18,13 +18,13 @@
 use std::process::Command;
 use std::str::FromStr;
 
+use carbide_libmlx_model::device::info::MlxDeviceInfo;
 use mac_address::MacAddress;
 use quick_xml::de::from_str;
 use serde::Deserialize;
 use tracing::{debug, warn};
 
 use crate::device::filters::DeviceFilter;
-use crate::device::info::MlxDeviceInfo;
 
 // DevicesXml represents the root XML structure
 // from mlxfwmanager output.
