+++ crates/libmlx/src/firmware/mod.rs
@@ -20,5 +20,4 @@ pub mod credentials;
 pub mod error;
 pub mod flasher;
 pub mod reset;
-pub mod result;
 pub mod source;
