+++ crates/libmlx/tests/variables/test_registry.rs
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
+use carbide_libmlx_model::device::info::MlxDeviceInfo;
 use libmlx::device::filters::{DeviceField, DeviceFilter, DeviceFilterSet, MatchMode};
-use libmlx::device::info::MlxDeviceInfo;
 use libmlx::variables::registry::MlxVariableRegistry;
 use libmlx::variables::spec::MlxVariableSpec;
 use libmlx::variables::variable::MlxConfigVariable;
